Move relevant files into the Form folder
